Updated WPA2 handshake to new hashcat format
Hashcat used to use the hccapx binary format for WPA2 handshakes
https://hashcat.net/wiki/doku.php?id=hccapx

They have now deprecated that format and moved to a different one
https://hashcat.net/forum/thread-10253.html and
https://github.com/hashcat/hashcat/issues/1816

This commit removed the code for generating the hccapx's and implements the new
format. It maintains an ability to write the hash to a file specified by the
mana_wpaout config directive.

But because the new format is now a string rather than binary, it will also
output it as a string.